moves subclasses to own folder
