+++ Gemfile
@@ -24,4 +24,5 @@ group :development, :test do
   gem "pry-byebug"
   gem "webmock"
   gem "simplecov"
+  gem "climate_control"
 end
