+++ env.example
@@ -1,5 +1,9 @@
 BUNDLE_RUBYGEMS__PKG__GITHUB__COM="your_github_personal_access_token_with_read_packages"
 
+# Setting this to test "turns" on all of the CLI subcommands. Setting this to a
+# subcommand limits the CLI to that subcommand
+APP_NAME="test"
+
 ALMA_API_KEY='your_alma_api_key'
 ALMA_API_HOST='https://api-na.hosted.exlibrisgroup.com'
