+++ spec/aim/hathifiles/updater_spec.rb
